+++ RELEASE-NOTES.md
@@ -38,6 +38,7 @@ Initial version with:
 
 ## [0.8.0](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.7.0...v0.8.0) (2022-07-12)
 * Contacts > Insert into a list
+* Contacts > Deduplicate list
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -47,6 +48,7 @@ Initial version with:
 
 ### New Features
 * Contacts > Insert into a list
+* Contacts > Deduplicate list
 
 ### Bug Fixes